tests/integration/testdata/buildcmd/CDK/cdk-example-multiple-stacks-01/requirements.txt
aws-cdk.core
aws-cdk.aws_sam
aws-cdk.aws_lambda
aws-cdk.aws_apigateway
